Refactoring
* renamed repositories to better readability
