+++ app/src/main/java/com/academy/android/domain/interactors/GetFilterParametersUseCase.kt
@@ -1,7 +1,7 @@
 package com.academy.android.domain.interactors
 
-import com.academy.android.data.repositories.VideosRepository
 import com.academy.android.domain.models.FilterParameters
+import com.academy.android.domain.repositories.VideosRepository
 import javax.inject.Inject
 
 class GetFilterParametersUseCase @Inject constructor(
