+++ app/src/main/java/com/academy/android/domain/interactors/GetFilteredVideosUseCase.kt
@@ -1,10 +1,10 @@
 package com.academy.android.domain.interactors
 
-import com.academy.android.data.repositories.VideosRepository
 import com.academy.android.domain.models.City
 import com.academy.android.domain.models.CourseLevel
 import com.academy.android.domain.models.CourseYear
 import com.academy.android.domain.models.Video
+import com.academy.android.domain.repositories.VideosRepository
 import javax.inject.Inject
 
 class GetFilteredVideosUseCase @Inject constructor(
