+++ app/src/main/java/com/academy/android/domain/interactors/GetLikesCountForChatIdUseCase.kt
@@ -1,11 +1,11 @@
 package com.academy.android.domain.interactors
 
-import com.academy.android.data.repositories.NewsRepositorySource
+import com.academy.android.domain.repositories.NewsRepository
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
 
 class GetLikesCountForChatIdUseCase @Inject constructor(
-    private val newsRepository: NewsRepositorySource
+    private val newsRepository: NewsRepository
 ) {
     operator fun invoke(chatId: Long): Flow<Int> =
         newsRepository.getLikesCountForId(chatId)
