+++ app/src/main/java/com/academy/android/domain/interactors/GetMessagesCountForChatIdUseCase.kt
@@ -1,11 +1,11 @@
 package com.academy.android.domain.interactors
 
-import com.academy.android.data.repositories.ChatsRepositorySource
+import com.academy.android.domain.repositories.ChatsRepository
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
 
 class GetMessagesCountForChatIdUseCase @Inject constructor(
-    private val chatsRepository: ChatsRepositorySource
+    private val chatsRepository: ChatsRepository
 ) {
     operator fun invoke(chatId: Long): Flow<Int> =
         chatsRepository.getMessagesCountForId(chatId)
