+++ app/src/main/java/com/academy/android/domain/interactors/GetPassedNewsUseCase.kt
@@ -1,14 +1,14 @@
 package com.academy.android.domain.interactors
 
-import com.academy.android.data.repositories.NewsRepositorySource
 import com.academy.android.domain.models.News
+import com.academy.android.domain.repositories.NewsRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
-import java.util.*
+import java.util.Date
 import javax.inject.Inject
 
 class GetPassedNewsUseCase @Inject constructor(
-    private val newsRepository: NewsRepositorySource
+    private val newsRepository: NewsRepository
 ) {
     operator fun invoke(date: Date = Date()): Flow<List<News>> {
         return newsRepository.newsList.map { list ->
