+++ app/src/main/java/com/academy/android/domain/repositories/ChatsRepository.kt
@@ -1,0 +1,8 @@
+package com.academy.android.domain.repositories
+
+import kotlinx.coroutines.flow.Flow
+
+
+interface ChatsRepository {
+    fun getMessagesCountForId(chatId: Long): Flow<Int>
+}
