+++ app/src/main/java/com/academy/android/domain/repositories/ProfileRepository.kt
@@ -1,0 +1,16 @@
+package com.academy.android.domain.repositories
+
+import com.academy.android.domain.models.Profile
+import kotlinx.coroutines.flow.StateFlow
+
+
+interface ProfileRepository{
+    val profileData: StateFlow<Profile>
+
+    fun saveProfileData(
+        profile: Profile
+    )
+
+    fun getProfileData(): Profile
+}
+
